apps/claude-runner/pyproject.toml
[project]
name = "ichor-claude-runner"
version = "0.0.0"
description = "Ichor — local Win11 FastAPI :8765, subprocesses claude -p headless via Max 20x"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.118.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "structlog>=25.1.0",
    "anyio>=4.7.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",  # CF Access JWT verify
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ichor_claude_runner"]
